+++ app/src/main/java/edu/ucne/composedemo/presentation/Ticket/TicketEvent.kt
@@ -1,7 +1,7 @@
 package edu.ucne.composedemo.presentation.Ticket
 
 sealed interface TicketEvent{
-    data class TicketChange(val ticketId: Double?): TicketEvent
+    data class TicketChange(val ticketId: Int?): TicketEvent
     data class ClienteChange(val cliente:Double?): TicketEvent
     data class EncargadoChange(val idEncargado:Int?): TicketEvent
     data class AsuntoChange(val asunto: String): TicketEvent
